scripts/ml_requirements.txt
# ML Pipeline Requirements
# For YOLOv8 object detection pipeline

ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0
python-dotenv>=1.0.0
